Refactored BucketListBase
