+++ src/bucket/BucketBase.cpp
@@ -9,7 +9,6 @@
 #include "bucket/BucketBase.h"
 #include "bucket/BucketIndex.h"
 #include "bucket/BucketInputIterator.h"
-#include "bucket/BucketList.h"
 #include "bucket/BucketListSnapshot.h"
 #include "bucket/BucketManager.h"
 #include "bucket/BucketOutputIterator.h"
