+++ src/bucket/BucketListBase.cpp
@@ -1,16 +1,13 @@
-// Copyright 2014 Stellar Development Foundation and contributors. Licensed
+// Copyright 2024 Stellar Development Foundation and contributors. Licensed
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "BucketList.h"
-#include "bucket/BucketIndexImpl.h"
+#include "bucket/BucketListBase.h"
 #include "bucket/BucketInputIterator.h"
 #include "bucket/BucketManager.h"
-#include "bucket/BucketSnapshot.h"
 #include "bucket/HotArchiveBucket.h"
 #include "bucket/LiveBucket.h"
 #include "crypto/SHA.h"
-#include "ledger/LedgerManager.h"
 #include "ledger/LedgerTxn.h"
 #include "main/Application.h"
 #include "util/GlobalChecks.h"
@@ -18,19 +15,11 @@
 #include "util/ProtocolVersion.h"
 #include "util/types.h"
 
-#include "medida/counter.h"
-
 #include <Tracy.hpp>
 #include <fmt/format.h>
 
 namespace stellar
 {
-
-template <> BucketListDepth BucketListBase<LiveBucket>::kNumLevels = 11;
-
-// TODO: This is an arbitrary number. Do some analysis and pick a better value
-// or make this a configurable network config.
-template <> BucketListDepth BucketListBase<HotArchiveBucket>::kNumLevels = 9;
 
 template <typename BucketT>
 BucketLevel<BucketT>::BucketLevel(uint32_t i)
@@ -692,209 +681,6 @@ BucketListBase<BucketT>::addBatchInternal(Application& app, uint32_t currLedger,
     }
 }
 
-void
-HotArchiveBucketList::addBatch(Application& app, uint32_t currLedger,
-                               uint32_t currLedgerProtocol,
-                               std::vector<LedgerEntry> const& archiveEntries,
-                               std::vector<LedgerKey> const& restoredEntries,
-                               std::vector<LedgerKey> const& deletedEntries)
-{
-    ZoneScoped;
-    releaseAssertOrThrow(protocolVersionStartsFrom(
-        currLedgerProtocol,
-        BucketBase::FIRST_PROTOCOL_SUPPORTING_PERSISTENT_EVICTION));
-    addBatchInternal(app, currLedger, currLedgerProtocol, archiveEntries,
-                     restoredEntries, deletedEntries);
-}
-
-void
-LiveBucketList::addBatch(Application& app, uint32_t currLedger,
-                         uint32_t currLedgerProtocol,
-                         std::vector<LedgerEntry> const& initEntries,
-                         std::vector<LedgerEntry> const& liveEntries,
-                         std::vector<LedgerKey> const& deadEntries)
-{
-    ZoneScoped;
-    addBatchInternal(app, currLedger, currLedgerProtocol, initEntries,
-                     liveEntries, deadEntries);
-}
-
-BucketEntryCounters
-LiveBucketList::sumBucketEntryCounters() const
-{
-    BucketEntryCounters counters;
-    for (auto const& lev : mLevels)
-    {
-        for (auto const& b : {lev.getCurr(), lev.getSnap()})
-        {
-            if (b->isIndexed())
-            {
-                auto c = b->getBucketEntryCounters();
-                counters += c;
-            }
-        }
-    }
-    return counters;
-}
-
-void
-LiveBucketList::updateStartingEvictionIterator(EvictionIterator& iter,
-                                               uint32_t firstScanLevel,
-                                               uint32_t ledgerSeq)
-{
-    // Check if an upgrade has changed the starting scan level to below the
-    // current iterator level
-    if (iter.bucketListLevel < firstScanLevel)
-    {
-        // Reset iterator to the new minimum level
-        iter.bucketFileOffset = 0;
-        iter.isCurrBucket = true;
-        iter.bucketListLevel = firstScanLevel;
-    }
-
-    // Whenever a Bucket changes (spills or receives an incoming spill), the
-    // iterator offset in that bucket is invalidated. After scanning, we
-    // must write the iterator to the BucketList then close the ledger.
-    // Bucket spills occur on ledger close after we've already written the
-    // iterator, so the iterator may be invalidated. Because of this, we
-    // must check if the Bucket the iterator currently points to changed on
-    // the previous ledger, indicating the current iterator is invalid.
-    if (iter.isCurrBucket)
-    {
-        // Check if bucket received an incoming spill
-        releaseAssert(iter.bucketListLevel != 0);
-        if (BucketListBase::levelShouldSpill(ledgerSeq - 1,
-                                             iter.bucketListLevel - 1))
-        {
-            // If Bucket changed, reset to start of bucket
-            iter.bucketFileOffset = 0;
-        }
-    }
-    else
-    {
-        if (BucketListBase::levelShouldSpill(ledgerSeq - 1,
-                                             iter.bucketListLevel))
-        {
-            // If Bucket changed, reset to start of bucket
-            iter.bucketFileOffset = 0;
-        }
-    }
-}
-
-bool
-LiveBucketList::updateEvictionIterAndRecordStats(
-    EvictionIterator& iter, EvictionIterator startIter,
-    uint32_t configFirstScanLevel, uint32_t ledgerSeq,
-    std::shared_ptr<EvictionStatistics> stats, EvictionCounters& counters)
-{
-    releaseAssert(stats);
-
-    // If we reached eof in curr bucket, start scanning snap.
-    // Last level has no snap so cycle back to the initial level.
-    if (iter.isCurrBucket && iter.bucketListLevel != kNumLevels - 1)
-    {
-        iter.isCurrBucket = false;
-        iter.bucketFileOffset = 0;
-    }
-    else
-    {
-        // If we reached eof in snap, move to next level
-        ++iter.bucketListLevel;
-        iter.isCurrBucket = true;
-        iter.bucketFileOffset = 0;
-
-        // If we have scanned the last level, cycle back to initial
-        // level
-        if (iter.bucketListLevel == kNumLevels)
-        {
-            iter.bucketListLevel = configFirstScanLevel;
-
-            // Record then reset metrics at beginning of new eviction cycle
-            stats->submitMetricsAndRestartCycle(ledgerSeq, counters);
-        }
-    }
-
-    // If we are back to the bucket we started at, break
-    if (iter.bucketListLevel == startIter.bucketListLevel &&
-        iter.isCurrBucket == startIter.isCurrBucket)
-    {
-        return true;
-    }
-
-    return false;
-}
-
-void
-LiveBucketList::checkIfEvictionScanIsStuck(EvictionIterator const& evictionIter,
-                                           uint32_t scanSize,
-                                           std::shared_ptr<LiveBucket const> b,
-                                           EvictionCounters& counters)
-{
-    // Check to see if we can finish scanning the new bucket before it
-    // receives an update
-    uint64_t period = bucketUpdatePeriod(evictionIter.bucketListLevel,
-                                         evictionIter.isCurrBucket);
-    if (period * scanSize < b->getSize())
-    {
-        CLOG_WARNING(Bucket,
-                     "Bucket too large for current eviction scan size.");
-        counters.incompleteBucketScan.inc();
-    }
-}
-
-// To avoid noisy data, only count metrics that encompass a complete
-// eviction cycle. If a node joins the network mid cycle, metrics will be
-// nullopt and be initialized at the start of the next cycle.
-void
-LiveBucketList::scanForEvictionLegacy(Application& app, AbstractLedgerTxn& ltx,
-                                      uint32_t ledgerSeq,
-                                      EvictionCounters& counters,
-                                      std::shared_ptr<EvictionStatistics> stats)
-{
-    releaseAssert(stats);
-
-    auto getBucketFromIter = [&levels = mLevels](EvictionIterator const& iter) {
-        auto& level = levels.at(iter.bucketListLevel);
-        return iter.isCurrBucket ? level.getCurr() : level.getSnap();
-    };
-
-    auto const& networkConfig =
-        app.getLedgerManager().getSorobanNetworkConfig();
-    auto const firstScanLevel =
-        networkConfig.stateArchivalSettings().startingEvictionScanLevel;
-    auto evictionIter = networkConfig.evictionIterator();
-    auto scanSize = networkConfig.stateArchivalSettings().evictionScanSize;
-    auto maxEntriesToEvict =
-        networkConfig.stateArchivalSettings().maxEntriesToArchive;
-
-    updateStartingEvictionIterator(evictionIter, firstScanLevel, ledgerSeq);
-
-    auto startIter = evictionIter;
-    auto b = getBucketFromIter(evictionIter);
-
-    while (b->scanForEvictionLegacy(
-               ltx, evictionIter, scanSize, maxEntriesToEvict, ledgerSeq,
-               counters.entriesEvicted, counters.bytesScannedForEviction,
-               stats) == Loop::INCOMPLETE)
-    {
-
-        if (updateEvictionIterAndRecordStats(evictionIter, startIter,
-                                             firstScanLevel, ledgerSeq, stats,
-                                             counters))
-        {
-            break;
-        }
-
-        b = getBucketFromIter(evictionIter);
-        checkIfEvictionScanIsStuck(
-            evictionIter,
-            networkConfig.stateArchivalSettings().evictionScanSize, b,
-            counters);
-    }
-
-    networkConfig.updateEvictionIterator(ltx, evictionIter);
-}
-
 template <typename BucketT>
 void
 BucketListBase<BucketT>::restartMerges(Application& app,
@@ -981,4 +767,18 @@ template class BucketListBase<LiveBucket>;
 template class BucketListBase<HotArchiveBucket>;
 template class BucketLevel<LiveBucket>;
 template class BucketLevel<HotArchiveBucket>;
+
+template void BucketListBase<HotArchiveBucket>::addBatchInternal<
+    std::vector<LedgerEntry>, std::vector<LedgerKey>, std::vector<LedgerKey>>(
+    Application& app, uint32_t currLedger, uint32_t currLedgerProtocol,
+    std::vector<LedgerEntry> const& archiveEntries,
+    std::vector<LedgerKey> const& restoredEntries,
+    std::vector<LedgerKey> const& deletedEntries);
+
+template void BucketListBase<LiveBucket>::addBatchInternal<
+    std::vector<LedgerEntry>, std::vector<LedgerEntry>, std::vector<LedgerKey>>(
+    Application& app, uint32_t currLedger, uint32_t currLedgerProtocol,
+    std::vector<LedgerEntry> const& initEntries,
+    std::vector<LedgerEntry> const& liveEntries,
+    std::vector<LedgerKey> const& deadEntries);
 }
