+++ src/bucket/test/BucketListTests.cpp
@@ -11,12 +11,13 @@
 // else.
 #include "util/asio.h"
 #include "bucket/BucketInputIterator.h"
-#include "bucket/BucketList.h"
 #include "bucket/BucketListSnapshot.h"
 #include "bucket/BucketManager.h"
 #include "bucket/BucketOutputIterator.h"
 #include "bucket/HotArchiveBucket.h"
+#include "bucket/HotArchiveBucketList.h"
 #include "bucket/LiveBucket.h"
+#include "bucket/LiveBucketList.h"
 #include "bucket/test/BucketTestUtils.h"
 #include "crypto/Hex.h"
 #include "ledger/LedgerTypeUtils.h"
