+++ src/catchup/ApplyBufferedLedgersWork.cpp
@@ -3,8 +3,8 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "catchup/ApplyBufferedLedgersWork.h"
-#include "bucket/BucketList.h"
 #include "bucket/BucketManager.h"
+#include "bucket/LiveBucketList.h"
 #include "catchup/ApplyLedgerWork.h"
 #include "crypto/Hex.h"
 #include "ledger/LedgerManager.h"
